+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Gestion des points
 - Modale pour afficher qui a buzzé
+- Authentification via Google
 - Nettoyage du code
 
 ## 1.0.0 (2022-07-23)